+++ packages/dash-table/tests/cypress/src/DashTable.ts
@@ -19,90 +19,94 @@ const getSelector = (state: State) => {
     }
 };
 
-export default class DashTable {
-    static getCell(row: number, column: number, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr td.column-${column}`).eq(row);
+export class DashTableHelper {
+    constructor(private readonly id) {
+
     }
 
-    static getCellById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr td[data-dash-column="${column}"]`).eq(row);
+    public getCell(row: number, column: number, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr td.column-${column}`).eq(row);
     }
 
-    static getFilter(column: number, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-filter.column-${column}`);
+    public getCellById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr td[data-dash-column="${column}"]`).eq(row);
     }
 
-    static getFilterById(column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-filter[data-dash-column="${column}"]`);
+    public getFilter(column: number, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-filter.column-${column}`);
     }
 
-    static getHeader(row: number, column: number, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header.column-${column}`).eq(row);
+    public getFilterById(column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-filter[data-dash-column="${column}"]`);
     }
 
-    static getHeaderById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"]`).eq(row);
+    public getHeader(row: number, column: number, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header.column-${column}`).eq(row);
     }
 
-    static focusCell(row: number, column: number) {
+    public getHeaderById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"]`).eq(row);
+    }
+
+    public focusCell(row: number, column: number) {
         // somehow we need to scrollIntoView AFTER click, or it doesn't
         // work right. Why?
         return this.getCell(row, column).click().scrollIntoView();
     }
 
-    static focusCellById(row: number, column: string) {
+    public focusCellById(row: number, column: string) {
         return this.getCellById(row, column).click().scrollIntoView();
     }
 
-    static clearColumnById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--clear`).eq(row).click();
+    public clearColumnById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--clear`).eq(row).click();
     }
 
-    static deleteColumnById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--delete`).eq(row).click();
+    public deleteColumnById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--delete`).eq(row).click();
     }
 
-    static hideColumnById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--hide`).eq(row).click();
+    public hideColumnById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--hide`).eq(row).click();
     }
 
-    static getSelectColumnById(row: number, column: string, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--select input`).eq(row);
+    public getSelectColumnById(row: number, column: string, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr th.dash-header[data-dash-column="${column}"] .column-header--select input`).eq(row);
     }
 
-    static selectColumnById(row: number, column: string) {
-        return DashTable.getSelectColumnById(row, column).click();
+    public selectColumnById(row: number, column: string) {
+        return this.getSelectColumnById(row, column).click();
     }
 
-    static getDelete(row: number, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr td.dash-delete-cell`).eq(row);
+    public getDelete(row: number, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr td.dash-delete-cell`).eq(row);
     }
 
-    static getSelect(row: number, editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody tr td.dash-select-cell`).eq(row);
+    public getSelect(row: number, editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody tr td.dash-select-cell`).eq(row);
     }
 
-    static getActiveCell(editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody td.focused`);
+    public getActiveCell(editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody td.focused`);
     }
 
-    static getSelectedCells(editable: State = State.Ready) {
-        return cy.get(`#table ${getSelector(editable)} tbody td.cell--selected`);
+    public getSelectedCells(editable: State = State.Ready) {
+        return cy.get(`#${this.id} ${getSelector(editable)} tbody td.cell--selected`);
     }
 
-    static scrollToTop() {
+    public scrollToTop() {
         cy.get(`.cell.cell-1-1.dash-fixed-content`).invoke(`outerHeight`).then(height => {
             cy.scrollTo(0, -height);
         });
     }
 
-    static scrollToBottom() {
+    public scrollToBottom() {
         cy.get(`.cell.cell-1-1.dash-fixed-content`).invoke(`outerHeight`).then(height => {
             cy.scrollTo(0, height);
         });
     }
 
-    static getCellInLastRowOfColumn(column: number) {
+    public getCellInLastRowOfColumn(column: number) {
         const cellInLastRow = cy.get(`td.dash-cell.column-${column}`).last().then(elem => {
             const lastRow = elem ? elem.attr(`data-dash-row`) : undefined;
             return lastRow ? cy.get(`td.dash-cell.column-${column}[data-dash-row="${lastRow}"`) : undefined;
@@ -110,13 +114,15 @@ export default class DashTable {
         return cellInLastRow;
     }
 
-    static getCellFromDataDash(row: number, column: number) {
+    public getCellFromDataDash(row: number, column: number) {
         return cy.get(`td.column-${column}[data-dash-row="${row}"]`);
     }
 
-    static toggleScroll(toggled: boolean) {
+    public toggleScroll(toggled: boolean) {
         cy.get('.row-1').then($el => {
             $el[0].style.overflow = toggled ? '' : 'unset';
         });
     }
 }
+
+export default new DashTableHelper('table');